kvs: move wait_version api into kvsroot
Problem: The kvs wait_version API is mostly a set of convenience
functions on the kvsroot.  The API is quite small and exists as
a separate file mostly due to the legacy implementation of KVS
"watchers" in the KVS.

Move the wait_version API functions into kvsroot, which ends up
only being 3 new convenience functions.  Move over the kvs wait_version
unit tests into the kvsroot ones.  Remove the old wait_version API
files and test files.
